Merge commit 'c98d68cd56338ec57695b3c35114fbde88d4bde8' into eclair-plus-aosp

* commit 'c98d68cd56338ec57695b3c35114fbde88d4bde8':
  Move SMS pdu creation fully into the central Phone process.

+++ telephony/java/com/android/internal/telephony/gsm/SmsMessage.java
@@ -337,7 +337,7 @@ public class SmsMessage extends SmsMessageBase{
      *         Returns null on encode error.
      */
     public static SubmitPdu getSubmitPdu(String scAddress,
-            String destinationAddress, short destinationPort, byte[] data,
+            String destinationAddress, int destinationPort, byte[] data,
             boolean statusReportRequested) {
 
         SmsHeader.PortAddrs portAddrs = new SmsHeader.PortAddrs();
